remove unused components/css
