--- src/components/Global/PoolCard/PoolCardSkeleton/PoolCardSkeleton.module.css
@@ -1,73 +1,0 @@
-.skeleton_wrapper {
-    width: 268px;
-    height: 125px;
-    background: var(--dark2);
-    position: relative;
-    overflow: hidden;
-    padding: 8px;
-}
-
-.skeleton_content {
-    width: 100%;
-
-    align-items: center;
-
-    height: 100%;
-
-    display: grid;
-    grid-template-columns: 0.9fr 1fr;
-    gap: 14px;
-}
-
-.left_side,
-.right_side {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    height: 100%;
-}
-
-.right_side {
-    /* justify-content: flex-end; */
-    align-items: flex-end;
-}
-.tokens {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 8px;
-}
-
-.token {
-    width: 30px;
-    height: 30px;
-    border-radius: 50%;
-    background: var(--dark1);
-}
-
-.amount {
-    background: var(--dark1);
-    width: 100%;
-    height: 26px;
-}
-
-.right_item1 {
-    background: var(--dark1);
-    height: 20px;
-    width: 100%;
-}
-.right_item2 {
-    background: var(--dark1);
-    height: 15px;
-    width: 90%;
-}
-.right_item3 {
-    background: var(--dark1);
-    height: 20px;
-    width: 80%;
-}
-.right_item4 {
-    background: var(--dark1);
-    height: 15px;
-    width: 70%;
-}
