--- src/components/Global/Skeletons/Shimmer.tsx
@@ -1,7 +1,0 @@
-export default function Shimmer() {
-    return (
-        <div className='shimmer-wrapper'>
-            <div className='shimmer'></div>
-        </div>
-    );
-}
