--- src/components/Global/Toggle/Toggle2.module.css
@@ -1,94 +1,0 @@
-/* .toggle_switch {
-    position: relative;
-    display: inline-block;
-    width: 36px;
-    height: 20px;
-    margin-top: -3px;
-   
-}
-.toggle_switch input[type='checkbox'] {
-    display: none;
-} */
-
-/* .toggle_switch .switch {
-    position: absolute;
-    cursor: pointer;
-
-    border-radius: 25px;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    transition: background-color var(--animation-speed) ease-in-out;
-
-    border: double 1px transparent;
-
-    background-image: linear-gradient(var(--dark2), var(--dark2)), var(--title-gradient);
-    background-origin: border-box;
-    background-clip: content-box, border-box;
-} */
-/* .toggle_switch .switch::before {
-    position: absolute;
-    content: '';
-
-    width: 16px;
-    height: 16px;
-
-    border-radius: 50%;
-    transition: transform var(--animation-speed) ease-in-out;
-
-    border: double 2px transparent;
-    border-radius: 50%;
-    background-image: linear-gradient(var(--text2), var(--text2)),
-        var(--title-gradient);
-    background-origin: border-box;
-    background-clip: content-box, border-box;
-}
-.toggle_switch input[type='checkbox']:checked + .switch::before {
-    transform: translateX(18px);
-    background-color: var(--text2);
-}
-.toggle_switch input[type='checkbox']:checked + .switch {
-    background: var(--title-gradient);
-} */
-
-.switch {
-    width: 36px;
-    height: 20px;
-    background: var(--dark3);
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    border-radius: 18px;
-    border: 1px solid var(--accent1);
-    cursor: pointer;
-}
-
-.switch:focus-visible {
-    box-shadow: 0px 0px 36px rgba(205, 193, 255, 0.2),
-        0px 0px 21px rgba(205, 193, 255, 0.2),
-        0px 0px 12px rgba(205, 193, 255, 0.2),
-        0px 0px 7px rgba(205, 193, 255, 0.2), 0px 0px 4px var(--accent5),
-        0px 0px 2px rgba(205, 193, 255, 0.2);
-}
-
-.switch[data-isOn='true'] {
-    justify-content: flex-end;
-    background-color: var(--accent1);
-}
-.switch[data-isOn='true'] .handle {
-    background: var(--text1);
-}
-
-.handle {
-    /* width: 36px; */
-    width: 18px;
-    height: 18px;
-    background: var(--text1);
-    border-radius: 40px;
-}
-
-.disabled {
-    pointer-events: none;
-    opacity: 0.4;
-}
