--- src/components/Swap/DenominationSwitch/DenominationSwitch.module.css
@@ -1,48 +1,0 @@
-.denomination_switch {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-    gap: 0.5rem;
-    min-height: 33px;
-    font-weight: 300;
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    text-align: right;
-    color: var(--text2);
-}
-
-.denomination_switch button {
-    cursor: pointer;
-    background: var(--dark2);
-    min-width: 1.5rem;
-    /* padding: 0.1rem; */
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    text-align: center;
-    color: var(--text2);
-    outline: none;
-    border: solid 0.2px transparent;
-    border-radius: 50px;
-    padding: 3px 8px;
-    transition: all var(--animation-speed) ease-in-out;
-}
-
-/* .denomination_switch button:hover{
-    background: var(--title-gradient);
-    transition: all var(--animation-speed) ease-in-out;
-   
-} */
-.denomination_switch .non_active_button:hover {
-    /* transition: all 0.4s ease-in-out; */
-    box-shadow: 0px 0px 36px rgba(205, 193, 255, 0.2),
-        0px 0px 21px rgba(205, 193, 255, 0.2),
-        0px 0px 12px rgba(205, 193, 255, 0.2),
-        0px 0px 7px rgba(205, 193, 255, 0.2), 0px 0px 4px var(--accent5),
-        0px 0px 2px rgba(205, 193, 255, 0.2);
-}
-
-.denomination_switch .active_button {
-    background: transparent;
-    border: 1px solid var(--accent1);
-    transition: box-shadow var(--animation-speed) ease-in-out;
-}
