--- src/components/TopRanges/TopRange.module.css
@@ -1,82 +1,0 @@
-.main_container {
-    /* margin: 3rem 0; */
-    display: grid;
-    grid-template-columns: auto 50px;
-    align-items: center;
-    height: 40px;
-}
-.tokens_container {
-    display: none;
-}
-.row_container {
-    display: flex;
-    align-items: center;
-    width: 100%;
-}
-
-.row_container section {
-    flex: 0.118 0;
-    /* flex-flow: column; */
-    align-items: center;
-    padding: 2px 8px;
-    /* font-size: 10px; */
-}
-
-.menu_container {
-    display: flex;
-    width: 100%;
-    justify-content: flex-end;
-}
-
-.apr {
-    font-weight: 200;
-
-    font-size: 12px;
-    line-height: 15px;
-
-    background: var(--apr-gradient-primary);
-    -webkit-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    background-clip: text;
-
-    text-shadow: 0px 0px 10px rgba(21, 190, 111, 0.75);
-}
-
-/* Mobile Landscape Screen Sizes */
-@media only screen and (min-width: 580px) {
-    .tokens_container {
-        display: flex;
-    }
-    .main_container {
-        grid-template-columns: 50px auto 50px;
-    }
-}
-
-/* Desktop screen Sizes */
-@media only screen and (min-width: 768px) {
-}
-
-/* Large devices (laptops/desktops, 992px and up) */
-@media only screen and (min-width: 992px) {
-    .main_container {
-        grid-template-columns: 100px auto 50px;
-        /* background: blue; */
-    }
-}
-
-/* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1200px) {
-}
-
-/* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1500px) {
-    .main_container {
-        grid-template-columns: 100px auto 250px;
-    }
-}
-/* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1600px) {
-    .main_container {
-        grid-template-columns: 100px 1.1fr 0.6fr;
-    }
-}
