--- src/components/TopRanges/TopRangeCardHeader.module.css
@@ -1,99 +1,0 @@
-.main_container {
-    display: grid;
-    grid-template-columns: auto 50px;
-    align-items: center;
-
-    font-size: 10px;
-
-    padding-bottom: 4px;
-    border-bottom: 1px solid var(--dark2);
-}
-
-.token_logos,
-.pool {
-    display: none;
-}
-
-.row_container {
-    display: flex;
-    align-items: center;
-    width: 100%;
-    padding: 0 8px;
-}
-
-.row_container p {
-    flex: 0.3 0;
-    /* flex-flow: column; */
-    align-items: center;
-
-    padding: 2px 8px;
-    /* font-size: var(--body-size);
-  line-height: var(--body-lh); */
-    color: rgba(235, 235, 255, 0.4);
-}
-
-.wallet,
-.range_sing,
-.token {
-    display: none;
-}
-
-/* .type{
-  text-align: end;
-} */
-
-/* .price{
-  background: rgb(30, 30, 2);
-  text-align: center;
-} */
-
-/* Mobile Landscape Screen Sizes */
-@media only screen and (min-width: 580px) {
-    .main_container {
-        grid-template-columns: 50px auto 50px;
-    }
-
-    .token_logos {
-        display: flex;
-    }
-}
-
-/* Desktop screen Sizes */
-@media only screen and (min-width: 768px) {
-    .wallet {
-        display: flex;
-    }
-}
-/* Large devices (laptops/desktops, 992px and up) */
-@media only screen and (min-width: 992px) {
-    .main_container {
-        grid-template-columns: 100px auto 50px;
-    }
-}
-
-/* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1200px) {
-    .tokens,
-    .range {
-        display: none;
-    }
-
-    .token,
-    .range_sing {
-        display: flex;
-    }
-}
-
-/* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1500px) {
-    .main_container {
-        grid-template-columns: auto 250px;
-    }
-}
-
-/* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1600px) {
-    .main_container {
-        grid-template-columns: 100px 1fr 0.6fr;
-    }
-}
