--- src/components/TopRanges/TopRanges.module.css
@@ -1,3 +1,0 @@
-.container {
-    padding: 0 8px;
-}
