--- src/components/TopTokens/TopTokenHeader.module.css
@@ -1,95 +1,0 @@
-.main_container {
-    display: grid;
-    grid-template-columns: auto 50px;
-    align-items: center;
-
-    font-size: 10px;
-
-    padding-bottom: 4px;
-    border-bottom: 1px solid var(--dark2);
-}
-.token_logos,
-.pool {
-    display: none;
-}
-.row_container {
-    display: flex;
-    align-items: center;
-    width: 100%;
-}
-
-.row_container p {
-    flex: 0.4 0;
-    /* flex-flow: column; */
-    align-items: center;
-
-    padding: 2px 8px;
-    /* font-size: var(--body-size);
-    line-height: var(--body-lh); */
-    color: rgba(235, 235, 255, 0.4);
-}
-
-.pCursor {
-    cursor: pointer;
-}
-
-.token {
-    display: none;
-    cursor: pointer;
-}
-
-/* Mobile Landscape Screen Sizes */
-@media only screen and (min-width: 580px) {
-    .main_container {
-        grid-template-columns: 50px auto;
-    }
-    .token_logos {
-        display: flex;
-    }
-}
-
-/* Medium screen devices (768px and above) */
-@media only screen and (min-width: 768px) {
-    .token {
-        display: none;
-    }
-    .header_container {
-        grid-template-columns: 100px auto;
-    }
-
-    .row {
-        grid-template-columns: 100px auto repeat(2, 50px) 40px 30px;
-    }
-}
-
-/* Large devices (laptops/desktops, 992px and up) */
-@media only screen and (min-width: 992px) {
-    .main_container {
-        grid-template-columns: 100px auto;
-    }
-}
-
-/* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1200px) {
-    .token {
-        display: flex;
-    }
-
-    /* .token, .side{
-        display: flex;
-    } */
-}
-
-/* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1500px) {
-    .main_container {
-        grid-template-columns: 100px auto;
-    }
-}
-
-/* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1600px) {
-    .main_container {
-        grid-template-columns: 100px 1fr;
-    }
-}
