--- src/components/TopTokens/TopTokenRow.module.css
@@ -1,143 +1,0 @@
-.main_container {
-    /* margin: 3rem 0; */
-    display: grid;
-    grid-template-columns: auto 50px;
-    align-items: center;
-    height: 40px;
-    cursor: pointer;
-}
-.tokens_container {
-    display: none;
-}
-.row_container {
-    display: flex;
-    align-items: center;
-    width: 100%;
-}
-
-.row_container section {
-    flex: 0.25 0;
-    /* flex-flow: column; */
-    align-items: center;
-    padding: 2px 8px;
-    /* font-size: 10px; */
-}
-
-.menu_container {
-    display: flex;
-    width: 100%;
-    justify-content: flex-end;
-}
-
-.token_key {
-    font-weight: 200;
-    font-size: 13px;
-    line-height: 32px;
-    letter-spacing: -0.02em;
-    color: var(--text1);
-    /* text-align: right; */
-    padding-left: 7px;
-}
-
-.token_icon {
-    display: flex;
-    align-items: center;
-    flex-direction: row;
-    gap: 4px;
-}
-
-.display {
-    display: flex;
-    color: rgba(235, 235, 255, 0.4);
-}
-.displayName {
-    display: none;
-    color: rgba(235, 235, 255, 0.4);
-}
-
-.trade_button {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    border-radius: 6.25rem;
-    cursor: pointer;
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    color: var(--text2);
-    height: 23px;
-    padding: 5px 8px;
-    text-decoration: none;
-
-    z-index: 1;
-    background: var(--dark1);
-    border: 1px solid var(--dark3);
-    box-shadow: 2px 1000px 1px var(--dark2) inset;
-    transition: box-shadow var(--animation-speed) ease-in-out;
-    cursor: pointer;
-}
-
-/* Desktop screen Sizes */
-@media only screen and (min-width: 768px) {
-    .display {
-        font-size: 10px;
-        line-height: 13px;
-    }
-
-    .displayName {
-        display: none;
-        font-size: 10px;
-        line-height: 13px;
-        /* font-size: 10px;
-    line-height: 13px;  */
-    }
-}
-
-/* Desktop screen Sizes */
-@media only screen and (min-width: 1200px) {
-    .displayName {
-        display: flex;
-
-        font-size: 10px;
-        line-height: 13px;
-    }
-}
-
-/* Mobile Landscape Screen Sizes */
-@media only screen and (min-width: 580px) {
-    .tokens_container {
-        display: flex;
-    }
-    .main_container {
-        grid-template-columns: 50px auto;
-        /* background: red; */
-    }
-}
-
-/* Desktop screen Sizes */
-@media only screen and (min-width: 768px) {
-}
-
-/* Large devices (laptops/desktops, 992px and up) */
-@media only screen and (min-width: 992px) {
-    .main_container {
-        grid-template-columns: 100px auto;
-        /* background: blue; */
-    }
-}
-
-/* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1200px) {
-}
-
-/* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1500px) {
-    .main_container {
-        grid-template-columns: 100px auto;
-    }
-}
-/* Extra large devices (large laptops and desktops, 1200px and up) */
-@media only screen and (min-width: 1600px) {
-    .main_container {
-        grid-template-columns: 100px 1fr;
-    }
-}
