--- src/components/Trade/Edit/AmountAndCurrencyDisplay/AmountAndCurrencyDisplay.module.css
@@ -1,151 +1,0 @@
-/* Currency Quantity */
-
-.currency_quantity {
-    font-weight: 200;
-    font-size: 18px;
-    line-height: 22px;
-    color: white;
-    text-align: start;
-    width: 100%;
-    height: calc(1.5em + 0.75rem + 2px);
-    padding: 0.375rem 0.75rem;
-    border: none;
-    border-radius: 0.25rem;
-    outline: 0;
-    background-color: transparent;
-    background-clip: padding-box;
-
-    transition: border-color var(--animation-speed) ease-in-out,
-        box-shadow var(--animation-speed) ease-in-out;
-}
-.currency_quantity::placeholder {
-    color: var(--text3);
-    font-weight: 200;
-    font-size: 18px;
-    line-height: 22px;
-}
-
-/* Currency Display */
-
-/* swapbox */
-.swapbox {
-    display: flex;
-    overflow: auto;
-    flex-direction: column;
-    overflow-y: hidden;
-    color: white;
-}
-
-.modalSwapbox {
-    background-color: var(--dark2);
-}
-
-.direction {
-    padding: 12px;
-}
-.direction,
-.swapbox_bottom {
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    color: var(--text2);
-}
-
-.swapbox_top {
-    border-radius: var(--border-radius);
-    background-color: var(--dark2);
-    display: flex;
-    justify-content: space-between;
-    flex-direction: row;
-    align-items: center;
-    /* padding: 0.5rem 0; */
-    min-height: 40px;
-    /* height: 40px; */
-}
-
-.swap_input {
-    width: 70%;
-}
-
-.token_select {
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    width: 30%;
-    cursor: pointer;
-}
-
-.token_select img {
-    max-width: 100%;
-}
-
-.token_list_text {
-    font-size: var(--header2-size);
-    line-height: var(--header2-lh);
-    color: var(--text1);
-    font-weight: 200;
-}
-
-/* bottom info */
-.swapbox_bottom {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-evenly;
-    align-items: center;
-}
-
-.token_select_button {
-    cursor: pointer;
-}
-
-.buy_token {
-    display: none;
-}
-
-.sellAllButton {
-    cursor: pointer;
-    transition: all var(--animation-speed) ease-out;
-
-    color: green;
-}
-
-.sellAllButton:hover {
-    color: hotpink;
-}
-.toggle_container {
-    margin-bottom: 10px;
-    margin-left: 10px;
-}
-
-.surplus_toggle {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-}
-
-/* modal */
-
-.pool_modal {
-    max-height: 400px;
-    overflow-y: scroll;
-    overflow-x: hidden;
-    transition: all var(--animation-speed)
-        cubic-bezier(0.175, 0.885, 0.32, 1.275);
-}
-
-.pool_modal::-webkit-scrollbar {
-    display: none;
-}
-
-.search_input input {
-    width: 100%;
-    background: var(--dark2);
-
-    text-align: center;
-    color: var(--text2);
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    padding: 1rem 0;
-    border-radius: var(--border-radius);
-    border: none;
-    outline: none;
-}
