--- src/components/Trade/Edit/CurrencyDisplayContainer/CurrencyDisplayContainer.module.css
@@ -1,5 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-}
