--- src/components/Trade/Edit/EditButton/EditButton.module.css
@@ -1,3 +1,0 @@
-.button_container {
-    padding: 8px 32px 0px;
-}
