--- src/components/Trade/Edit/EditHeader/EditHeader.module.css
@@ -1,7 +1,0 @@
-.title {
-    font-weight: 300;
-    font-size: 24px;
-    line-height: 30px;
-
-    color: var(--text1);
-}
