--- src/components/Trade/Edit/EditPriceInfo/EditPriceInfo.module.css
@@ -1,114 +1,0 @@
-.price_info_container {
-    padding: 0 5%;
-    transition: all var(--animation-speed) ease;
-    margin-bottom: 1rem;
-    transition: height var(--animation-speed) ease;
-}
-
-.price_info_content {
-    transition: all var(--animation-speed) ease-in-out;
-    background: var(--dark2);
-    margin-top: 10px;
-    padding: 8px 16px;
-    border-radius: var(--border-radius);
-    color: var(--text2);
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-}
-.price_info_row {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0.2rem 0;
-}
-
-.current {
-    margin-left: 170px;
-}
-
-.apr {
-    display: flex;
-    justify-content: center;
-    margin: 10px 0;
-    font-size: var(--header2-size);
-    line-height: var(--header2-lh);
-    color: var(--other-green);
-}
-
-.advanced_table_display table {
-    color: var(--text2);
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    border-collapse: collapse;
-    width: 100%;
-}
-
-.advanced_table_display th {
-    text-align: end;
-}
-
-.advanced_table_display td,
-.advanced_table_display th {
-    padding: 10px 0;
-}
-.advanced_table_display td:nth-child(2),
-.advanced_table_display td:nth-child(3) {
-    text-align: end;
-}
-
-/* 
-Max width before this PARTICULAR table gets nasty
-This query will take effect for any screen smaller than 760px
-and also iPads specifically.
-*/
-@media only screen and (max-width: 760px),
-    (min-device-width: 768px) and (max-device-width: 1024px) {
-    .price_info_container {
-        padding: 0;
-    }
-
-    .advanced_table_display table {
-        width: 100%;
-    }
-
-    /* Force table to not be like tables anymore */
-    .advanced_table_display table,
-    .advanced_table_display thead,
-    .advanced_table_display tbody,
-    .advanced_table_display th,
-    .advanced_table_display td,
-    .advanced_table_display tr {
-        display: block;
-    }
-
-    /* Hide table headers (but not display: none;, for accessibility) */
-    .advanced_table_display thead tr {
-        position: absolute;
-        top: -9999px;
-        left: -9999px;
-    }
-
-    .advanced_table_display td {
-        /* Behave  like a "row" */
-        border: none;
-        border-bottom: 1px solid var(--text2);
-        position: relative;
-        padding-left: 50%;
-    }
-
-    .advanced_table_display td:before {
-        /* Now like a table header */
-        position: absolute;
-        /* Top/left values mimic padding */
-        top: 6px;
-        left: 6px;
-        width: 45%;
-        padding-right: 10px;
-        white-space: nowrap;
-        /* Label the data */
-        content: attr(data-column);
-
-        font-weight: bold;
-    }
-}
