--- src/components/Trade/Edit/EditPriceInput/EditPriceInput.module.css
@@ -1,57 +1,0 @@
-.minMax_container {
-    display: flex;
-    flex-direction: column;
-    width: 48%;
-    justify-content: center;
-    align-items: center;
-}
-
-.title,
-.percentage {
-    color: var(--text2);
-    font-size: var(--body-size);
-    line-height: var(--body-lh);
-    padding: 5px;
-}
-
-.percentage {
-    color: var(--accent5);
-    font-size: var(--header2-size);
-    line-height: var(--header2-lh);
-}
-
-.price_input_container {
-    display: grid;
-    grid-template-columns: 30px 1fr 30px;
-    color: var(--text1);
-    font-size: var(--header1-size);
-    line-height: var(--header1-lh);
-    text-align: center;
-
-    background: var(--dark2);
-    border-radius: var(--border-radius);
-    padding: 0.8rem 5px;
-}
-
-.price_quantity {
-    position: relative;
-    overflow: hidden;
-    width: 100%;
-    padding: 0;
-    color: var(--text1);
-    text-align: center;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    border: none;
-    outline: none;
-    background-color: transparent;
-    flex: 1 1 auto;
-    -webkit-appearance: textfield;
-    -moz-appearance: textfield;
-    appearance: textfield;
-}
-.currency_quantity::placeholder {
-    color: var(--text3);
-    font-size: var(--header1-size);
-    line-height: var(--header1-lh);
-}
