--- src/utils/hooks/useWindowDimensions.ts
@@ -1,32 +1,0 @@
-import { useState, useEffect } from 'react';
-
-// Define general type for useWindowSize hook, which includes width and height
-
-function getWindowDimensions() {
-    const { innerWidth: width, innerHeight: height } = window;
-    return {
-        width,
-        height,
-    };
-}
-
-function useWindowDimensions() {
-    const [windowDimensions, setWindowDimensions] = useState(
-        getWindowDimensions(),
-    );
-
-    useEffect(() => {
-        function handleResize() {
-            setWindowDimensions(getWindowDimensions());
-        }
-
-        window.addEventListener('resize', handleResize);
-        return () => window.removeEventListener('resize', handleResize);
-    }, []);
-
-    return windowDimensions;
-}
-
-export default useWindowDimensions;
-
-// Taken from https://usehooks.com/useWindowSize/: -Jr
